fix(api): restore cache and auth compliance [CODX-P1-API-311]
